Rename `add_meta_tags` -> `tag`
Almost half the API methods begin with `add`. This is pretty explicitly
a logging library, so saying `add` is redundant. The `meta` is never
really explained either, so we drop it and just use the verb `tag`,
which carries the intent well enough. After all, we're adding tags to an
experiment, like the learning rate.